requirements.txt
langchain==0.2.17
langchain-core==0.2.43
langchain-experimental==0.0.65
langchain-text-splitters==0.2.4
langchain-community==0.2.19
langchain-aws==0.1.18
langchain-cohere==0.1.9
langchain-google-genai==1.0.10
langchain-openai==0.1.25
langgraph==0.2.46
langsmith==0.1.142
anthropic==0.39.0
google-generativeai==0.7.2
faiss-cpu==1.9.0
chromadb==0.4.24
unstructured
pypdf
docx2txt
python-decouple
openpyxl
azure-core
azure-identity
azure-storage-blob
azure-ai-documentintelligence
duckduckgo_search
# crewai==0.55.2
# crewai-tools==0.13.1
xmltodict
biopython
html2text
requests-html
accelerate
transformers==4.46.2
bitsandbytes==0.42.0
sentence_transformers==3.3.0
peft==0.13.2
cohere
arxiv
# llama-cpp-python